backend/requirements.txt
flask>=2.0.3
flask-cors>=3.0.10
joblib>=1.1.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.27.1
python-dotenv>=1.0.0
reportlab>=3.6.12
matplotlib>=3.7.0
flask-sqlalchemy>=2.5.1
flask-bcrypt>=0.7.1
flask-jwt-extended>=4.3.1
psycopg2>=2.9.10